requirements.txt
aiohttp
azure-ai-formrecognizer
azure-storage-blob
azure-ai-documentintelligence==1.0.0b2
docx2pdf==0.1.7
python-dotenv
pypandoc==1.6.4
pandas==1.3.5
quart